+++ spec/rails_app/app/components/grid/associations.rb
@@ -1,0 +1,7 @@
+class Grid::Associations < Netzke::Basepack::Grid
+  def configure(c)
+    super
+    c.model = 'Book'
+    c.columns = [:title, :author__first_name, :author__prize_count]
+  end
+end