+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.12.1 - WIP
 * Fix filtering by association when associtaion method is not string
 * Fix decimal columns in advanced search form
+* Fix showing 0-value in association shown by integer
 
 # 0.12.0 - 2015-03-18
 * ExtJS 5